falta ajustes no bonus

+++ client.c
@@ -6,7 +6,7 @@
 /*   By: dida-sil <dida-sil@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/23 13:01:12 by dida-sil          #+#    #+#             */
-/*   Updated: 2022/08/23 14:27:22 by dida-sil         ###   ########.fr       */
+/*   Updated: 2022/08/24 19:17:28 by dida-sil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int	main(int argc, char *argv[])
 	pid_t	n_pid;
 
 	signal(SIGUSR1, ft_return);
-	signal(SIGUSR2, ft_return);
+	signal(SIGUSR2, ft_failure);
 	if ((argc != 3) || (!ft_isalldigit(argv[1])))
 	{
 		ft_printf("Incorrect\n");
@@ -25,10 +25,15 @@ int	main(int argc, char *argv[])
 		exit(EXIT_FAILURE);
 	}
 	n_pid = ft_atoi(argv[1]);
-	while (*argv[2])
+	while (1)
+	{
 		ft_atobin(*argv[2]++, n_pid);
-	ft_atobin(0, n_pid);
-	kill(getpid(), SIGUSR2);
+		if (*argv[2] == '\0')
+		{
+			ft_atobin(0, n_pid);
+			kill(getpid(), SIGUSR2);
+		}
+	}
 }
 
 void	ft_atobin(char c, pid_t n_pid)
@@ -50,22 +55,23 @@ int	ft_isalldigit(char	*str)
 {
 	while (*str)
 	{
-		if (!ft_isdigit(*str++))
+		if (!ft_isdigit(*str))
 			return (0);
+		str++;
 	}
 	return (1);
 }
 
-void	ft_return(int signum)
+void	ft_return(int sig)
 {
-	if (signum == SIGUSR1)
-	{
-		ft_printf ("Success\n");
-		exit(EXIT_SUCCESS);
-	}
-	else
-	{
-		ft_printf ("Failed\n");
-		exit(EXIT_FAILURE);
-	}
+	(void)sig;
+	ft_printf("%s", "Message Sent to Server\n");
+	exit(EXIT_SUCCESS);
+}
+
+void	ft_failure(int sig)
+{
+	(void)sig;
+	ft_printf("%s", "Message Not Sent to Server\n");
+	exit(EXIT_FAILURE);
 }
